requirements.txt
alembic==1.11.1
anyio==4.0.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
dnspython==2.4.2
ecdsa==0.18.0
email-validator==2.0.0.post2
fastapi==0.95.2
greenlet==2.0.2
h11==0.14.0
httpcore==0.18.0
httptools==0.6.0
httpx==0.25.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
orjson==3.9.7
passlib==1.7.4
pyasn1==0.5.0
pycparser==2.21
pydantic==1.10.12
python-decouple==3.8
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0.1
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.41
sqlalchemy2-stubs==0.0.2a35
sqlmodel==0.0.8
starlette==0.27.0
typing_extensions==4.8.0
ujson==5.8.0
urllib3==2.0.5
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.20.0
websockets==11.0.3
